Separate MapperContext into interface and implementation
This allows us to hide the details of how contexts are mutated, which
are irrelevant to end users.

In the generated code, Mapper implementations use DefaultMapperContext,
but DAO implementations only need MapperContext because they don't
mutate it.
